refactor: calendar as alpine component
Implements the calendar view as Alpine.js component.
In the component itself, the `init` and `render` methods become superfluous
as the rendering is done by alpine and the view is defined declaratively
in the html code.

The store now needs to use native events via `document.dispatchEvent`
instead of using jQuery for alpine to pick them up.